requirements.txt
pylibmc==1.2.3
django-pylibmc-sasl==0.2.4
Django==1.5.1
Mezzanine==1.4.7
Pillow==2.0.0
South==0.8.1
bleach==1.2.2
boto==2.9.5
dj-database-url==0.2.1
django-storages==1.1.8
filebrowser-safe==0.2.27
grappelli-safe==0.2.21
gunicorn==0.17.4
html5lib==0.95
mezzanine-flexipage==0.1.3
psycopg2==2.5
pytz==2013b
wsgiref==0.1.2
geopy==0.95.1
newrelic==1.12.0.56
django-localflavor==1.0
